requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
Pillow==9.0.0
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.5
cryptography==41.0.2
decorator==5.1.1
docarray==0.13.17
docx2txt==0.8
python-docx
entrypoints==0.4
exceptiongroup==1.1.2
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-api-python-client==2.83.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
h11==0.12.0
httpcore==0.13.7
httplib2==0.22.0
httpx==0.18.2
idna==3.4
importlib-metadata==6.8.0
itranslate==0.1.2
Jinja2==3.1.2
joblib==1.3.1
jsonlines==3.0.0
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
logzero==1.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdtable==0.2
mdurl==0.1.2
multidict==6.0.4
numpy==1.25.1
openai==0.27.4
packaging==23.1
pandas==1.5.3
pdfminer.six==20220319
protobuf==3.20.3
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.1.0
PyPDF2==3.0.1
python-dateutil==2.8.2
pytube==12.1.3
pytz==2023.3
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
rfc3986==1.5.0
rich==13.4.2
rpds-py==0.8.11
rsa==4.9
scipy==1.11.1
semver==3.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
streamlit==1.25.0
streamlit-chat
streamlit-pills
tiktoken==0.3.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing_extensions==4.7.1
tzlocal==4.0.1
uritemplate==4.1.1
urllib3==1.26.16
validators==0.20.0
Wand==0.6.11
yarl==1.9.2
zipp==3.16.2
